Impliment first few API calls and testing
Currently all testing is done against the main brawlhalla server, in the
future this may be changed to a mock server.

The current system simply creates objects in the layout the server
provides, but this will eventually be changed into a single lazy loaded
player object. This will allow all information to be accessed much more
naturally.